+++ api/homesectionone.php
@@ -7,9 +7,9 @@ require('./config/dbconfig.php');
 
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-
-    if($_POST['method'] === 'post') {
-        $fileName = preg_replace("/[^a-zA-Z0-9.]/", "", $_FILES["file"]["name"]);
+    if ($_POST['method'] === 'post') {
+        $fileName = preg_replace("/[^a-zA-Z0-9.]/", "", $_POST["fileName"]);
+        $fileName = time().'_'.$fileName;
         $fileTmpLoc = $_FILES["file"]["tmp_name"]; // File in the PHP tmp folder
         $fileType = $_FILES["file"]["type"]; // The type of file it is
         $fileSize = $_FILES["file"]["size"]; // File size in bytes
@@ -27,26 +27,30 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         } else {
             echo "move_uploaded_file function failed";
         }
-    } else if($_POST['method'] === 'delete') {
+    } elseif ($_POST['method'] === 'delete') {
         $id = $_POST['id'];
-        $query = "DELETE FROM tbl_event WHERE id=".$id;
-    
-        if(mysqli_query($DB, $query)) {
-            echo json_encode(array("response"=> true));
+        $sql = "SELECT `filepath` FROM `tbl_event` WHERE id=".$id;
+        if ($row = mysqli_query($DB, $sql)) {
+            $result = mysqli_fetch_assoc($row);
+            $resultSplit = explode("/", $result["filepath"]);
+            $fileName = $resultSplit[count($resultSplit) - 1];
+            unlink('./uploads/homesectionone/'.$fileName);
+            $query = "DELETE FROM tbl_event WHERE id=".$id;
+            if (mysqli_query($DB, $query)) {
+                echo json_encode(array("response"=> true));
+            } else {
+                echo json_encode(array("response"=> false));
+            }
         } else {
             echo json_encode(array("response"=> false));
         }
-    
     }
 }
 
 if ($_SERVER['REQUEST_METHOD'] == 'GET') {
-
     $queryString = "select * from tbl_event order by id desc";
 
     $row = mysqli_query($DB, $queryString);
 
     echo json_encode(mysqli_fetch_all($row));
 }
-
-
